replace initialisation of shared_ptr from using new to using std::make_shared. This to remove memory leak (still some to be replaced)

+++ EAFramework/src/EA.cpp
@@ -36,6 +36,6 @@ Individual::Ptr EA::getIndividual(size_t index) const
 }
 
 void EA::fill_ind_to_eval(std::vector<int> &ind_to_eval){
-    for(int i = 0; i < population.size(); i++)
+    for(size_t i = 0; i < population.size(); i++)
         ind_to_eval.push_back(i);
 }
